requirements-dev.txt
-r requirements.txt
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-httpx==0.35.0
ruff==0.11.5
